pyproject.toml
[project]
name = "pa053"
version = "0.1.0"
description = "PA053 homework"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "grpcio>=1.71.0",
    "grpcio-tools>=1.71.0",
    "numpy>=2.2.4",
]

[project.scripts]
hello = "cat README.md"
